Dense varint-based serialization for DirStats
The directory stats DB needs to be uploaded after any change
to the backed-up data, so we want it to be as small as possible.
We want to be able to work with large backups

This database needs to contain an entry for every backed-up folder.
Because the B2 API is pretty limited, we need this data to be able to
diff only the folders whose contents changed, instead of having to
list every file (billed as 1 request per every 1000 files listed!).

We don't use serde because we want to squeeze the data in as small
a space possible, even if serialization takes slightly longer.
This format pulls a couple simple tricks to avoid storing information
we don't need or that we can just compute back during deserialization.

This is about 25% smaller than a plain binary format.