Retry failing HTTP requests harder and more consistently
Requests have been observed to fail before being able to get an
actual response from the server.
This happens much more often that it should, which I suspect is a
bug somewhere in the HTTP stack. We now handle this and retry.

We also won't give up on requests that keep failing anymore,
the exponential backoff will reach a ceiling of 3.2s and keep trying.
The previous behavior of eventually dropping failing requests
was Really Dumb™, with the fast moving output you could have thought
your backup was successful when it's actually missing several files...