Simplify signal handling
The signal caught flag is permanent now, we don't reset it.
Previously we would handle signals in the top-level code and wait
for tasks for finish before returning.
It's simpler if tasks can also check for signals and return early
instead of trying to have the top level code send messages.