DirDB no longer saves the root dir's name/name_hash
This didn't make sense (and could trigger a harmless debug assert).
The contents of the backup folder are relative to the root folder,
but the root can't have a name relative to itself.

This could cause a mismatch between what the remote DirDB's root
thinks its name is, and whatever folder name we have locally.